moving getDefaultAppDeps into runners/utils
