+++ src/index.ts
@@ -1,4 +1,5 @@
-import { getAppAsync, getDefaultAppDependenciesAsync } from './app';
+import { getAppAsync } from './app';
+import { getDefaultAppDependenciesAsync } from './runners/utils';
 import { defaultHttpServiceConfig } from './config';
 import { logger } from './logger';
 import { providerUtils } from './utils/provider_utils';
